scripts/research/requirements.txt
requests>=2.31.0
feedparser>=6.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
